Third variation on observer pattern
Use of generics: any type can be passed at each update
Problem: too much generic! The abstract observer is not interested in all kind of data, but only the data that can consume. For example, a player can consume a .mp3 file, a .avi file, but is not interested in a .doc.